fix(pipeline): --resume reuses run dirs and checkpoints; outputs/latest symlink (UB-06, T1.7)
Every invocation minted a fresh timestamp, so _find_latest_checkpoint()
always scanned an empty outputs/<new-ts>/checkpoints dir and the trainer's
complete resume machinery was dead across restarts.

- --resume <run_id> reuses outputs/<run_id> and logs/<run_id>; a missing
  run dir raises an actionable FileNotFoundError listing available IDs (R4).
- outputs/latest symlink updated atomically in Pipeline.__init__; platforms
  without symlink support warn instead of crashing.
- validate() returns plain Python floats: numpy scalars in the checkpoint's
  metric history were rejected by torch.load(weights_only=True) (R8),
  which the red test surfaced as the second half of the defect.
- test_resume.py: two-phase subprocess run — epoch 1, then --resume with
  NUM_EPOCHS=2 → history length 2 in both folds' metrics JSON, no second
  run dir, outputs/latest resolves to the run; unknown run id fails
  actionably.

+++ codes/unified_training.py
@@ -524,15 +524,18 @@ class UnifiedTrainer:
                 pbar.set_postfix({'loss': f'{loss.item():.4f}'})
         
         avg_loss = total_loss / len(self.val_loader)
-        avg_inference_time = np.mean(inference_times)
-        
+        # Plain Python floats: numpy scalars in the metric history would be
+        # rejected by the checkpoint's torch.load(weights_only=True) (R8),
+        # killing resume (UB-06).
+        avg_inference_time = float(np.mean(inference_times))
+
         # Compute epoch-level metrics
         ious_per_class = self.val_iou_metric.compute()
         # Remove NaNs (classes not present in target) before computing mean
         valid_ious = ious_per_class[~torch.isnan(ious_per_class)]
         mean_iou = valid_ious.mean().item() if valid_ious.numel() > 0 else 0.0
-        
-        mean_dice = np.mean(all_dice_scores)
+
+        mean_dice = float(np.mean(all_dice_scores))
         
         return avg_loss, mean_iou, mean_dice, avg_inference_time
